requirements.txt
selenium
beautifulsoup4
pytest
python-dateutil
fastapi
uvicorn
sqlalchemy
apscheduler
requests
dateparser 